third_party/blink/web_tests/external/wpt/html/semantics/forms/the-input-element/click-user-gesture-expected.txt

This is a testharness.js-based test.
[FAIL] input[type=color] click() consumes user activation
  assert_false: expected false got true
[FAIL] input[type=date] click() consumes user activation
  assert_false: expected false got true
[FAIL] input[type=datetime-local] click() consumes user activation
  assert_false: expected false got true
[FAIL] input[type=file] click() consumes user activation
  assert_false: expected false got true
[FAIL] input[type=month] click() consumes user activation
  assert_false: expected false got true
[FAIL] input[type=time] click() consumes user activation
  assert_false: expected false got true
[FAIL] input[type=week] click() consumes user activation
  assert_false: expected false got true
Harness: the test ran to completion.
